LPD-37555 - Move the test into a better file
